+++ docs/beta/en_US/changelog.md
@@ -1,3 +1,7 @@
+> 2024-08-15
+  + V3.0.1
+  + Add a configuration to limit the number of Blink's servers calls (Every minutes, every 5 min., etc).
+
 > 2024-04-07
   + V3.0.0
   + This version is compliant with Jeedom 4.4 (without JQuery)